ci(ubuntu): adds GitHub Action for build and test on ubuntu
This PR adds a GitHub Action for building and testing dice on the latest
ubuntu image. We combine the typical setup workflows for OCaml and Rust
projects, otherwise following the README.

Notably, in order to get the build to work, I also remove a test from
the test suite; after confirming that the feature being tested is no
longer suggested, I think this is appropriate.

Note: I also ran into problems getting this to work on macOS (a rsdd's
.so file is not being linked properly) and Windows (many, many
problems); I imagine I can tackle this in a follow-up PR!